@@ -4,7 +4,7 @@ import 'package:flutter/services.dart';
 
 class WifiMulticastLock {
   static const MethodChannel _channel = const MethodChannel('wifi_multicast_lock');
-  static WifiMulticastLock _lock;
+  static WifiMulticastLock? _lock;
 
   factory WifiMulticastLock() => _lock ??= WifiMulticastLock._();
 
@@ -14,7 +14,7 @@ class WifiMulticastLock {
 
   Future<void> release() async => await _channel.invokeMethod('release');
 
-  Future<bool> isHeld() async => await _channel.invokeMethod('isHeld');
+  Future<bool?> isHeld() async => await _channel.invokeMethod('isHeld');
 
   Future<void> setReferenceCounted(bool refCounted) async => await _channel.invokeMethod('setReferenceCounted', refCounted);
 }
